feat(Passwords): Ensure cannot reset password when token is expired

+++ roleplay-api/app/Controllers/Http/PasswordsController.ts
@@ -1,5 +1,6 @@
 import Mail from '@ioc:Adonis/Addons/Mail'
 import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
+import TokenExpired from 'App/Exceptions/TokenExpiredException'
 import User from 'App/Models/User'
 import ForgotPassword from 'App/Validators/ForgotPasswordValidator'
 import ResetPassword from 'App/Validators/ResetPasswordValidator'
@@ -40,6 +41,8 @@ export default class PasswordsController {
       .preload('tokens')
       .firstOrFail()
 
+    const tokenAge = Math.abs(userByToken.tokens[0].createdAt.diffNow('hours').hours)
+    if (tokenAge > 2) throw new TokenExpired()
     userByToken.password = password
     await userByToken.save()
     await userByToken.tokens[0].delete()
